+++ packages/honeycomb-opentelemetry-web/smoke-tests/test_helpers/utilities.bash
@@ -86,6 +86,24 @@ assert_equal() {
 	fi
 }
 
+# Fail and display details if the command that bats `run` executes exits
+# with a non-zero status. This function reads the $status and $output
+# that `run` sets.
+# Arguments: none
+assert_command_finished_successfully() {
+	if [[ "$status" -ne 0 ]]; then
+		{
+			echo
+			echo "-- 💥 command failed 💥 --"
+			echo "status : $status"
+			echo "output : $output"
+			echo "--"
+			echo
+		} >&2 # output error to STDERR
+		return 1
+	fi
+}
+
 # Fail and display details if the actual value is empty.
 # Arguments: $1 - actual result
 assert_not_empty() {